core/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qc-core"
version = "0.1.0"
description = "QC Core – GPU-capable ingest + QC engine for calcium imaging"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "grpcio>=1.63.0",
    "grpcio-tools>=1.63.0",
    "protobuf>=4.25.0",
    "websockets>=12.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "lz4>=4.3.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0",
    "pyarrow>=16.0.0",
    "starlette>=0.37.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "pytest-cov>=5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["qc_core*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
